refactor: modal에서 render관련 view요소 분리
